+++ js/generatePhotoGallery/index.js
@@ -1,0 +1,3 @@
+import { generatePhotoGallery } from './generate-photo-gallery.js';
+
+export { generatePhotoGallery };
