--- js/generatePhotoGallery/main.js
@@ -1,3 +1,0 @@
-import { generatePhotoGallery } from './data.js';
-
-generatePhotoGallery(25);
