+++ js/main.js
@@ -1,0 +1,3 @@
+import { generatePhotoGallery } from './generatePhotoGallery/index.js';
+
+generatePhotoGallery(25);
