+++ js/util.js
@@ -1,0 +1,33 @@
+const getRandomInteger = (min, max) => {
+  const lower = Math.ceil(Math.min(min, max));
+  const upper = Math.floor(Math.max(min, max));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+function getRandomUnicueInteger(min, max) {
+  const previousValues = [];
+
+  return function () {
+    let currentValue = getRandomInteger(min, max);
+    if (previousValues.length >= (max - min + 1)) {
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+}
+
+const getUniqueInteger = () => {
+  let lastGeneratedInteger = 0;
+
+  return function () {
+    lastGeneratedInteger += 1;
+    return lastGeneratedInteger;
+  };
+};
+
+export { getRandomInteger, getRandomUnicueInteger, getUniqueInteger };
